+++ sdk/ai/azure-ai-resources/azure/ai/resources/_telemetry/logging_handler.py
@@ -9,7 +9,8 @@
 import logging
 import platform
 import traceback
-import sys
+import os
+from typing import Optional
 
 from opencensus.ext.azure.log_exporter import AzureLogHandler
 from opencensus.ext.azure.common import utils
@@ -21,16 +22,14 @@ from opencensus.ext.azure.common.protocol import (
 )
 from opencensus.trace import config_integration
 
-from azureml.telemetry import INSTRUMENTATION_KEY
-
-from azure.ai.ml._telemetry.logging_handler import in_jupyter_notebook, CustomDimensionsFilter
+from azure.ai.ml._telemetry.logging_handler import in_jupyter_notebook, CustomDimensionsFilter, INSTRUMENTATION_KEY
 from azure.ai.resources._version import VERSION
 
 
 USER_AGENT = "{}/{}".format("azure-ai-resources", VERSION)
 
 
-GEN_AI_INTERNAL_LOGGER_NAMESPACE = "azure.ai.resources._telemetry"
+AI_RESOURCES_INTERNAL_LOGGER_NAMESPACE = "azure.ai.resources._telemetry"
 
 test_subscriptions = [
     "b17253fa-f327-42d6-9686-f3e553e24763",
@@ -43,22 +42,21 @@ test_subscriptions = [
 
 # activate operation id tracking
 config_integration.trace_integrations(["logging"])
-logging.basicConfig(
-    format="%(asctime)s traceId=%(traceId)s spanId=%(spanId)s %(message)s"
-)
 
 
-class OpsLogger:
+class ActivityLogger:
     def __init__(self, name: str):
-        self.package_logger: logging.Logger = logging.getLogger(GEN_AI_INTERNAL_LOGGER_NAMESPACE + name)
+        self.package_logger: logging.Logger = logging.getLogger(AI_RESOURCES_INTERNAL_LOGGER_NAMESPACE + name)
         self.package_logger.propagate = False
         self.module_logger = logging.getLogger(name)
-        self.custom_dimensions = {}
+        self.custom_dimensions: dict = {}
 
-    def update_info(self, data: dict) -> None:
-        if "app_insights_handler" in data:
-            self.package_logger.addHandler(data.pop("app_insights_handler"))
-
+    def update_info(self, data: Optional[dict] = None) -> None:
+        if data and "app_insights_handler" in data:
+            handler = data.pop("app_insights_handler")
+        else:
+            handler = get_appinsights_log_handler(USER_AGENT)
+        self.package_logger.addHandler(handler)
 
 
 # cspell:ignore overriden
@@ -67,7 +65,6 @@ def get_appinsights_log_handler(
     *args,  # pylint: disable=unused-argument
     instrumentation_key=None,
     component_name=None,
-    enable_telemetry=True,
     **kwargs,
 ):
     """Enable the OpenCensus logging handler for specified logger and instrumentation key to send info to AppInsights.
@@ -80,8 +77,6 @@ def get_appinsights_log_handler(
     :paramtype instrumentation_key: str
     :keyword component_name: The component name.
     :paramtype component_name: str
-    :keyword enable_telemetry: Whether to enable telemetry. Will be overriden to False if not in a Jupyter Notebook.
-    :paramtype enable_telemetry: bool
     :keyword kwargs: Optional keyword arguments for adding additional information to messages.
     :paramtype kwargs: dict
     :return: The logging handler.
@@ -91,7 +86,8 @@ def get_appinsights_log_handler(
         if instrumentation_key is None:
             instrumentation_key = INSTRUMENTATION_KEY
 
-        if not in_jupyter_notebook() or not enable_telemetry:
+        enable_telemetry = os.getenv("AZURE_AI_RESOURCES_ENABLE_LOGGING", "True")
+        if not in_jupyter_notebook() or enable_telemetry == "False":
             return logging.NullHandler()
 
         if not user_agent or not user_agent.lower() == USER_AGENT.lower():
@@ -102,7 +98,7 @@ def get_appinsights_log_handler(
                 return logging.NullHandler()
 
         child_namespace = component_name or __name__
-        current_logger = logging.getLogger(GEN_AI_INTERNAL_LOGGER_NAMESPACE).getChild(child_namespace)
+        current_logger = logging.getLogger(AI_RESOURCES_INTERNAL_LOGGER_NAMESPACE).getChild(child_namespace)
         current_logger.propagate = False
         current_logger.setLevel(logging.CRITICAL)
 
@@ -160,8 +156,8 @@ class AzureGenAILogHandler(AzureLogHandler):
             "process": record.processName,
             "module": record.module,
             "level": record.levelname,
-            "operation_id": envelope.tags["gen.ai.operation.id"],
-            "operation_parent_id": envelope.tags["gen.ai.operation.parentId"],
+            "operation_id": envelope.tags.get("ai.resources.operation.id"),
+            "operation_parent_id": envelope.tags.get("ai.resources.operation.parentId"),
         }
         if hasattr(record, "custom_dimensions") and isinstance(record.custom_dimensions, dict):
             properties.update(record.custom_dimensions)
@@ -225,14 +221,11 @@ def create_envelope(instrumentation_key, record):
         tags=dict(utils.azure_monitor_context),
         time=utils.timestamp_to_iso_str(record.created),
     )
-    envelope.tags["gen.ai.operation.id"] = getattr(
+    envelope.tags["ai.resources.operation.id"] = getattr(
         record,
         "traceId",
         "00000000000000000000000000000000",
     )
-    envelope.tags["gen.ai.operation.parentId"] = "|{}.{}.".format(
-        envelope.tags["gen.ai.operation.id"],
-        getattr(record, "spanId", "0000000000000000"),
-    )
+    envelope.tags["ai.resources.operation.parentId"] = f"|{envelope.tags.get('ai.resources.operation.id')}.{getattr(record, 'spanId', '0000000000000000')}"
 
     return envelope