+++ sdk/communication/azure-communication-identity/azure/communication/identity/__init__.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-import warnings
+
 from ._communication_identity_client import CommunicationIdentityClient
 from ._generated.models import CommunicationTokenScope
 
@@ -38,14 +38,3 @@ __all__ = [
     "PhoneNumberProperties",
     "UnknownIdentifier",
 ]
-
-def __getattr__(name):
-    if name == 'MicrosoftBotProperties':
-        warnings.warn(f"{name} is deprecated and should not be used.", DeprecationWarning)
-        from ._shared.models  import _MicrosoftBotProperties
-        return _MicrosoftBotProperties
-    if name == 'MicrosoftBotIdentifier':
-        warnings.warn(f"{name} is deprecated and should not be used.", DeprecationWarning)
-        from ._shared.models  import _MicrosoftBotIdentifier
-        return _MicrosoftBotIdentifier
-    raise AttributeError(f"module 'azure.communication.identity' has no attribute {name}")