+++ sdk/healthcareapis/azure-mgmt-healthcareapis/generated_samples/operation_results_get.py
@@ -36,6 +36,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/healthcareapis/resource-manager/Microsoft.HealthcareApis/stable/2022-06-01/examples/OperationResultsGet.json
+# x-ms-original-file: specification/healthcareapis/resource-manager/Microsoft.HealthcareApis/stable/2023-11-01/examples/OperationResultsGet.json
 if __name__ == "__main__":
     main()