+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_builders/parallel.py
@@ -11,9 +11,19 @@ import re
 from enum import Enum
 from typing import Dict, List, Optional, Union
 
-from marshmallow import Schema
+from marshmallow import INCLUDE, Schema
+
+from azure.ai.ml._schema.core.fields import NestedField, UnionField
+from azure.ai.ml._schema.job.identity import AMLTokenIdentitySchema, ManagedIdentitySchema, UserIdentitySchema
+from azure.ai.ml.entities._credentials import (
+    AmlTokenConfiguration,
+    ManagedIdentityConfiguration,
+    UserIdentityConfiguration,
+    _BaseJobIdentityConfiguration,
+)
 
 from ..._schema import PathAwareSchema
+from ..._utils.utils import is_data_binding_expression
 from ...constants._common import ARM_ID_PREFIX
 from ...constants._component import NodeType
 from .._component.component import Component
@@ -75,6 +85,12 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
                            the partition keys will take effect.
                            The input(s) must be partitioned dataset(s),
                            and the partition_keys must be a subset of the keys of every input dataset for this to work.
+    :keyword identity: The identity that the command job will use while running on compute.
+    :paramtype identity: Optional[Union[
+        dict[str, str],
+        ~azure.ai.ml.entities.ManagedIdentityConfiguration,
+        ~azure.ai.ml.entities.AmlTokenConfiguration,
+        ~azure.ai.ml.entities.UserIdentityConfiguration]]
     :type partition_keys: List
     :param input_data: The input data
     :type input_data: str
@@ -117,6 +133,9 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
         mini_batch_size: Optional[int] = None,
         resources: Optional[JobResourceConfiguration] = None,
         environment_variables: Optional[Dict] = None,
+        identity: Optional[
+            Union[ManagedIdentityConfiguration, AmlTokenConfiguration, UserIdentityConfiguration]
+        ] = None,
         **kwargs,
     ) -> None:
         # validate init params are valid type
@@ -150,7 +169,11 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
 
         self._task = task
 
-        if mini_batch_size is not None and not isinstance(mini_batch_size, int):
+        if (
+            mini_batch_size is not None
+            and not isinstance(mini_batch_size, int)
+            and not is_data_binding_expression(mini_batch_size)
+        ):
             """Convert str to int."""  # pylint: disable=pointless-string-statement
             pattern = re.compile(r"^\d+([kKmMgG][bB])*$")
             if not pattern.match(mini_batch_size):
@@ -179,7 +202,7 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
         self.mini_batch_error_threshold = mini_batch_error_threshold
         self._resources = resources
         self.environment_variables = {} if environment_variables is None else environment_variables
-
+        self._identity = identity
         if isinstance(self.component, ParallelComponent):
             self.resources = self.resources or copy.deepcopy(self.component.resources)
             self.input_data = self.input_data or self.component.input_data
@@ -242,6 +265,42 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
         if isinstance(value, dict):
             value = JobResourceConfiguration(**value)
         self._resources = value
+
+    @property
+    def identity(
+        self,
+    ) -> Optional[Union[ManagedIdentityConfiguration, AmlTokenConfiguration, UserIdentityConfiguration]]:
+        """The identity that the job will use while running on compute.
+
+        :return: The identity that the job will use while running on compute.
+        :rtype: Optional[Union[~azure.ai.ml.ManagedIdentityConfiguration, ~azure.ai.ml.AmlTokenConfiguration,
+            ~azure.ai.ml.UserIdentityConfiguration]]
+        """
+        return self._identity
+
+    @identity.setter
+    def identity(
+        self,
+        value: Union[
+            Dict[str, str], ManagedIdentityConfiguration, AmlTokenConfiguration, UserIdentityConfiguration, None
+        ],
+    ) -> None:
+        """Sets the identity that the job will use while running on compute.
+
+        :param value: The identity that the job will use while running on compute.
+        :type value: Union[dict[str, str], ~azure.ai.ml.ManagedIdentityConfiguration,
+            ~azure.ai.ml.AmlTokenConfiguration, ~azure.ai.ml.UserIdentityConfiguration]
+        """
+        if isinstance(value, dict):
+            identity_schema = UnionField(
+                [
+                    NestedField(ManagedIdentitySchema, unknown=INCLUDE),
+                    NestedField(AMLTokenIdentitySchema, unknown=INCLUDE),
+                    NestedField(UserIdentitySchema, unknown=INCLUDE),
+                ]
+            )
+            value = identity_schema._deserialize(value=value, attr=None, data=None)
+        self._identity = value
 
     @property
     def component(self) -> Union[str, ParallelComponent]:
@@ -349,6 +408,7 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
             retry_settings=self.retry_settings,
             input_data=self.input_data,
             logging_level=self.logging_level,
+            identity=self.identity,
             max_concurrency_per_instance=self.max_concurrency_per_instance,
             error_threshold=self.error_threshold,
             mini_batch_error_threshold=self.mini_batch_error_threshold,
@@ -395,6 +455,7 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
                     "partition_keys": json.dumps(self.partition_keys)
                     if self.partition_keys is not None
                     else self.partition_keys,
+                    "identity": self.identity._to_dict() if self.identity else None,
                     "resources": get_rest_dict_for_node_attrs(self.resources),
                 }
             )
@@ -423,6 +484,9 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
 
         if "partition_keys" in obj and obj["partition_keys"]:
             obj["partition_keys"] = json.dumps(obj["partition_keys"])
+
+        if "identity" in obj and obj["identity"]:
+            obj["identity"] = _BaseJobIdentityConfiguration._load(obj["identity"])
         return obj
 
     def _build_inputs(self):
@@ -477,6 +541,7 @@ class Parallel(BaseNode, NodeWithGroupInputMixin):
             node._base_path = self.base_path
             node.resources = copy.deepcopy(self.resources)
             node.environment_variables = copy.deepcopy(self.environment_variables)
+            node.identity = copy.deepcopy(self.identity)
             return node
         raise Exception(
             f"Parallel can be called as a function only when referenced component is {type(Component)}, "