+++ tools/azure-sdk-tools/ci_tools/scenario/generation.py
@@ -91,7 +91,7 @@ def create_package_and_install(
 
             if not pre_download_disabled:
                 requirements = ParsedSetup.from_path(os.path.join(os.path.abspath(target_setup), "setup.py")).requires
-                azure_requirements = [req.split(";")[0] for req in requirements if req.startswith("azure")]
+                azure_requirements = [req.split(";")[0] for req in requirements if req.startswith("azure-")]
 
                 if azure_requirements:
                     logging.info(